Fixing all pylance errors + broken references + refactored 1DE file parser (removed IO conflict).
