+++ backend/api/one_de_routes.py
@@ -1,18 +1,18 @@
 from typing import Any
 from fastapi import APIRouter, UploadFile
-from logic.simulation_1de import Simulation_1de
+from logic.one_de_simulation import Simulation_1de
+
 
 router = APIRouter(
     prefix="/1d",
     tags=["1D Simulation"]
 )
+
+
 @router.post('/ProteinInfo/File', response_model=list[Any])
 async def fileGetProteinInfo(file: UploadFile) -> Any:
-    
+
     return Simulation_1de.fileGetProteinInfo(file)
-
-
-
 
 
 @router.post('/BatchFileProtein/Batch', response_model=list[list[Any]])
