+++ backend/api/two_de_routes.py
@@ -1,31 +1,25 @@
-import  re
-import numpy as np
 from typing import List, Dict, Any
 from fastapi import APIRouter, UploadFile, File
-from logic.simulation_2de import Simulation_2de
+from logic.two_de_simulation import Simulation_2de
+
+
 router = APIRouter(
     prefix="/2d",
     tags=["2D Simulation"]
 )
-# -------------------------------------------------------------------
-# Endpoints
-# -------------------------------------------------------------------
+
+
 @router.post("/parse-fasta")
 async def parse_fasta(files: List[UploadFile] = File(...)):
- 
-    color_palette = [
-        '#FF0000', '#00FF00', '#0000FF', '#FFFF00', '#FF00FF', '#00FFFF',
-        '#FFA500', '#800080', '#008000', '#FFC0CB', '#A52A2A', '#808080'
-    ]
     new_proteins = []
 
-    for i, file in enumerate(files):
+    for _, file in enumerate(files):
         content = await file.read()
         sequences = Simulation_2de.parse_fasta_content(content.decode("utf-8"))
         Simulation_2de.parse_fasta(sequences,new_proteins)
         
-
     return new_proteins
+
 
 @router.post("/simulate-ief")
 async def run_ief_simulation(data: Dict[str, Any]):
@@ -43,4 +37,3 @@ async def run_sds_simulation(data: Dict[str, Any]):
     acrylamide_percentage = data.get("acrylamidePercentage", 7.5)
     canvas_height = data.get("canvasHeight", 600)
     return Simulation_2de.simulate_sds(proteins, y_axis_mode, acrylamide_percentage, canvas_height)
-
