--- backend/tests/Electro2DTests/testdata/electrophoresis1dStandards.fasta
@@ -1,68 +1,0 @@
->pdb|6X1Q|D Chain D, Beta-galactosidase
-MITDSLAVVLQRRDWENPGVTQLNRLAAHPPFASWRNSEEARTDRPSQQLRSLNGEWRFAWFPAPEAVPE
-SWLECDLPEADTVVVPSNWQMHGYDAPIYTNVTYPITVNPPFVPTENPTGCYSLTFNVDESWLQEGQTRI
-IFDGVNSAFHLWCNGRWVGYGQDSRLPSEFDLSAFLRAGENRLAVMVLRWSDGSYLEDQDMWRMSGIFRD
-VSLLHKPTTQISDFHVATRFNDDFSRAVLEAEVQMCGELRDYLRVTVSLWQGETQVASGTAPFGGEIIDE
-RGGYADRVTLRLNVENPKLWSAEIPNLYRAVVELHTADGTLIEAEACDVGFRVVRIENGLLLLNGKPLLI
-RGVNRHEHHPLHGQVMDEQTMVQDILLMKQNNFNAVRCSHYPNHPLWYTLCDRYGLYVVDEANIETHGMV
-PMNRLTDDPRWLPAMSERVTRMVQRDRNHPSVIIWSLGNESGHGANHDALYRWIKSVDPSRPVQYEGGGA
-DTTATDIICPMYARVDEDQPFPAVPKWSIKKWLSLPGETRPLILCEYAHAMGNSLGGFAKYWQAFRQYPR
-LQGGFVWDWVDQSLIKYDENGNPWSAYGGDFGDTPNDRQFCMNGLVFADRTPHPALTEAKHQQQFFQFRL
-SGQTIEVTSEYLFRHSDNELLHWMVALDGKPLASGEVPLDVAPQGKQLIELPELPQPESAGQLWLTVRVV
-QPNATAWSEAGHISAWQQWRLAENLSVTLPAASHAIPHLTTSEMDFCIELGNKRWQFNRQSGFLSQMWIG
-DKKQLLTPLRDQFTRAPLDNDIGVSEATRIDPNAWVERWKAAGHYQAEAALLQCTADTLADAVLITTAHA
-WQHQGKTLFISRKTYRIDGSGQMAITVDVVVASDTPHPARIGLNCQLAQVAERVNWLGLGPQENYPDRLT
-AACFDRWDLPLSDMYTPYVFPSENGLRCGTRELNYGPHQWRGDFQFNISRYSQQQLMETSHRHLLHAEEG
-TWLNIDGFHMGIGGDDSWSPSVSAEFQLSAGRYHYQLVWCQ
-
->pdb|2PRI|A Chain A, GLYCOGEN PHOSPHORYLASE B
-SRPLSDQEKRKQISVRGLAGVENVTELKKNFNRHLHFTLVKDRNVATPRDYYFALAHTVRDHLVGRWIRT
-QQHYYEKDPKRIYYLSLEFYMGRTLQNTMVNLALENACDEATYQLGLDMEELEEIEEDAGLGNGGLGRLA
-ACFLDSMATLGLAAYGYGIRYEFGIFNQKICGGWQMEEADDWLRYGNPWEKARPEFTLPVHFYGRVEHTS
-QGAKWVDTQVVLAMPYDTPVPGYRNNVVNTMRLWSAKAPNDFNLKDFNVGGYIQAVLDRNLAENISRVLY
-PNDNFFEGKELRLKQEYFVVAATLQDIIRRFKSSKFGCRDPVRTNFDAFPDKVAIQLNDTHPSLAIPELM
-RVLVDLERLDWDKAWEVTVKTCAYTNHTVIPEALERWPVHLLETLLPRHLQIIYEINQRFLNRVAAAFPG
-DVDRLRRMSLVEEGAVKRINMAHLCIAGSHAVNGVARIHSEILKKTIFKDFYELEPHKFQNKTNGITPRR
-WLVLCNPGLAEIIAERIGEEYISDLDQLRKLLSYVDDEAFIRDVAKVKQENKLKFAAYLEREYKVHINPN
-SLFDVQVKRIHEYKRQLLNCLHVITLYNRIKKEPNKFVVPRTVMIGGKAAPGYHMAKMIIKLITAIGDVV
-NHDPVVGDRLRVIFLENYRVSLAEKVIPAADLSEQISTAGTEASGTGNMKFMLNGALTIGTMDGANVEMA
-EEAGEENFFIFGMRVEDVDRLDQRGYNAQEYYDRIPELRQIIEQLSSGFFSPKQPDLFKDIVNMLMHHDR
-FKVFADYEEYVKCQERVSALYKNPREWTRMVIRNIATSGKFSSDRTIAQYAREIWGVEPSRQRLPAPDEK
-IP
-
->pdb|4F5S|B Chain B, Serum albumin
-DTHKSEIAHRFKDLGEEHFKGLVLIAFSQYLQQCPFDEHVKLVNELTEFAKTCVADESHAGCEKSLHTLF
-GDELCKVASLRETYGDMADCCEKQEPERNECFLSHKDDSPDLPKLKPDPNTLCDEFKADEKKFWGKYLYE
-IARRHPYFYAPELLYYANKYNGVFQECCQAEDKGACLLPKIETMREKVLTSSARQRLRCASIQKFGERAL
-KAWSVARLSQKFPKAEFVEVTKLVTDLTKVHKECCHGDLLECADDRADLAKYICDNQDTISSKLKECCDK
-PLLEKSHCIAEVEKDAIPENLPPLTADFAEDKDVCKNYQEAKDAFLGSFLYEYSRRHPEYAVSVLLRLAK
-EYEATLEECCAKDDPHACYSTVFDKLKHLVDEPQNLIKQNCDQFEKLGEYGFQNALIVRYTRKVPQVSTP
-TLVEVSRSLGKVGTRCCTKPESERMPCTEDYLSLILNRLCVLHEKTPVSEKVTKCCTESLVNRRPCFSAL
-TPDETYVPKAFDEKLFTFHADICTLPDTEKQIKKQTALVELLKHKPKATEEQLKTVMENFVAFVDKCCAA
-DDKEACFAVEGPKLVVSTQTALA
-
->pdb|1JTI|AAA68882.1 ovalbumin [Gallus gallus]
-MDSISVTNAKFCFDVFNEMKVHHVNENILYCPLSILTALAMVYLGARGNTESQMKKVLHFDSITGAGSTT
-DSQCGSSEYVHNLFKELLSEITRPNATYSLEIADKLYVDKTFSVLPEYLSCARKFYTGGVEEVNFKTAAE
-EARQLINSWVEKETNGQIKDLLVSSSIDFGTTMVFINTIYFKGIWKIAFNTEDTREMPFSMTKEESKPVQ
-MMCMNNSFNVATLPAEKMKILELPYASGDLSMLVLLPDEVSGLERIEKTINFDKLREWTSTNAMAKKSMK
-VYLPRMKIEEKYNLTSILMALGMTDLFSRSANLTGISSVDNLMISDAVHGVFMEVNEEGTEATGSTGAIG
-NIKHSLELEEFRADHPFLFFIRYNPTNAILFFGRYWSP
-
->gb|NP_001344263.1| carbonic anhydrase 2 [Mus musculus]
-MSHHWGYSKHNGPENWHKDFPIANGDRQSPVDIDTATAQHDPALQPLLISYDKAASKSIVNNGHSFNVEF
-DDSQDNAVLKGGPLSDSYRLIQFHFHWGSSDGQGSEHTVNKKKYAAELHLVHWNTKYGDFGKAVQQPDGL
-AVLGIFLKIGPASQGLQKVLEALHSIKTKGKRAAFANFDPCSLLPGNLDYWTYPGSLTTPPLLECVTWIV
-LREPITVSSEQMSHFRTLNFNEEGDAEEAMVDNWRPAQPLKNRKIKASFK
-
->gb|AFP63821.1| pancreatic trypsin inhibitor [Musca domestica]
-MEILTRASLLHVLLLFMVNFSQNNFVQGAVGLYGPDRYEGLPETCLQPMDFGYCRAKLQRYYFDIRRMKC
-TMFYWGGCAGNDNNFKTMDECNDYCSAGYEQSNSIVPLKRERSENNPNSNTNSNARSSNFNSRGNTAVNT
-SKSTAPKFTIKSSPAPSYRDNASANNGTARANVKTFNINPPQKELKYPTATDDLEDEEYDE
-
->sp|Q6L6Q5.1|LYS_OSTED
-MSAVLVLALVLLSLTCVTDAISDACLTCICKQESYGCTQIGCRMDGRSLSCGYFQIKKSYWIDCGRLGSS
-WEACADDYNCAVRCVRAYMKKYIGKSGCTANCKNYARLHNGGPKGCTKPSTLTYWNAVKNQGCSINS
-
->gb|CAA01755.1| aprotinin [synthetic construct]
-MRPDFCLEPPYTGPCVARIIRYFYNAKAGLCQTFVYGGCRAKRNNFKSAEDCERTCGGA
