+++ backend/tests/data/e_coliK12.faa
